fix switched files src/resampling.jl <-> test/resampling
